migrate address and student module

+++ Modules/Address/Database/Migrations/2019_05_16_153550_create_identifcations_table.php
@@ -19,7 +19,7 @@ class CreateIdentifcationsTable extends Migration
             $table->string('identifcation_number')->nullable();
             $table->date('issue_date')->nullable();
             $table->string('issue_place')->nullable();
-            $table->unsignedInteger('identable_id');
+            $table->unsignedBigInteger('identable_id');
             $table->timestamps();
         });
     }
